Update master to AWS prototype

+++ LogiusWebApp/src/main/java/org/verapdf/crawler/core/validation/ValidationService.java
@@ -1,5 +1,8 @@
 package org.verapdf.crawler.core.validation;
 
+import com.adobe.xmp.XMPDateTime;
+import com.adobe.xmp.XMPDateTimeFactory;
+import com.adobe.xmp.XMPException;
 import io.dropwizard.hibernate.UnitOfWork;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,96 +13,101 @@ import org.verapdf.crawler.api.validation.error.ValidationError;
 import org.verapdf.crawler.db.DocumentDAO;
 import org.verapdf.crawler.db.ValidationErrorDAO;
 import org.verapdf.crawler.db.ValidationJobDAO;
+import org.verapdf.crawler.tools.AbstractService;
 
+import javax.xml.datatype.DatatypeConfigurationException;
+import javax.xml.datatype.DatatypeFactory;
+import javax.xml.datatype.XMLGregorianCalendar;
 import java.io.*;
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
-public class ValidationService implements Runnable {
+public class ValidationService extends AbstractService {
+
     private static final Logger logger = LoggerFactory.getLogger(ValidationService.class);
+
+    private static final String PROPERTY_NAME_MOD_DATE_XMP = "modDateXMP";
+	private static final String PROPERTY_NAME_MOD_DATE_INFO_DICT = "modDateInfoDict";
+
+	private static final long SLEEP_DURATION = 60*1000;
 
     private final ValidationJobDAO validationJobDAO;
     private final ValidationErrorDAO validationErrorDAO;
     private final DocumentDAO documentDAO;
     private final PDFValidator validator;
-    private boolean running = false;
-    private boolean isAborted = false;
-    private String stopReason;
+    private ValidationJob currentJob;
 
     public ValidationService(ValidationJobDAO validationJobDAO, ValidationErrorDAO validationErrorDAO, DocumentDAO documentDAO, PDFValidator validator) {
-        this.validationJobDAO = validationJobDAO;
+        super("ValidationService", SLEEP_DURATION);
+    	this.validationJobDAO = validationJobDAO;
         this.validationErrorDAO = validationErrorDAO;
         this.documentDAO = documentDAO;
         this.validator = validator;
     }
 
-    public boolean isRunning() {
-        return running;
+	public ValidationJob getCurrentJob() {
+		return currentJob;
+	}
+
+	private void setCurrentJob(ValidationJob currentJob) {
+    	synchronized (ValidationService.class) {
+			this.currentJob = currentJob;
+		}
+	}
+
+    public void abortCurrentJob() {
+		try {
+			logger.info("Aborting current job");
+			currentJob.setStatus(ValidationJob.Status.ABORTED);
+			validator.terminateValidation();
+		} catch (IOException e) {
+			logger.error("Can't terminate current job", e);
+		}
+	}
+
+	@Override
+	protected void onStart() throws InterruptedException, ValidationDeadlockException {
+		setCurrentJob(retrieveCurrentJob());
+		if (currentJob != null) {
+			try {
+				processStartedJob();
+			} catch (IOException e) {
+				saveErrorResult(e);
+			}
+		}
+	}
+
+	@Override
+	protected boolean onRepeat() throws ValidationDeadlockException, InterruptedException {
+		setCurrentJob(retrieveNextJob());
+		if (currentJob != null) {
+			logger.info("Validating " + currentJob.getId());
+			try {
+				validator.startValidation(currentJob);
+				processStartedJob();
+			} catch (IOException e) {
+				saveErrorResult(e);
+			}
+			return false;
+		}
+		return true;
+	}
+
+	private void processStartedJob() throws IOException, ValidationDeadlockException, InterruptedException {
+        VeraPDFValidationResult result = validator.getValidationResult(currentJob);
+        saveResult(result);
     }
 
-    public String getStopReason() {
-        return stopReason;
-    }
-
-    public void start() {
-        running = true;
-        stopReason = null;
-        new Thread(this, "Thread-ValidationService").start();
-    }
-
-    public void abort() throws IOException {
-        isAborted = true;
-        validator.terminateValidation();
-    }
-
-    @Override
-    public void run() {
-        logger.info("Validation service started");
-        try {
-            ValidationJob job = currentJob();
-            if (job != null) {
-                try {
-                    processStartedJob(job);
-                } catch (IOException e) {
-                    saveErrorResult(job, e);
-                }
-            }
-
-            while (running) {
-                job = nextJob();
-                if (job != null) {
-                    logger.info("Validating " + job.getDocument().getUrl());
-                    try {
-                        validator.startValidation(job);
-                        processStartedJob(job);
-                    } catch (IOException e) {
-                        saveErrorResult(job, e);
-                    }
-                    continue;
-                }
-
-                Thread.sleep(60 * 1000);
-            }
-        } catch (Throwable e) {
-            logger.error("Fatal error in validator, stopping validation service.", e);
-            this.stopReason = e.getMessage();
-        } finally {
-            running = false;
-        }
-    }
-
-    private void processStartedJob(ValidationJob job) throws IOException, ValidationDeadlockException, InterruptedException {
-        VeraPDFValidationResult result = validator.getValidationResult(job);
-        saveResult(job, result);
-    }
-
-    private void saveErrorResult(ValidationJob job, Throwable e) {
+    private void saveErrorResult(Throwable e) {
         VeraPDFValidationResult result = new VeraPDFValidationResult(e.getMessage());
-        saveResult(job, result);
+        saveResult(result);
     }
 
     @SuppressWarnings("WeakerAccess") // @UnitOfWork works only with public methods
     @UnitOfWork
-    public ValidationJob nextJob() {
+    public ValidationJob retrieveNextJob() {
+		logger.debug("Getting next job");
         ValidationJob job = validationJobDAO.next();
         if (job != null) {
             job.setStatus(ValidationJob.Status.IN_PROGRESS);
@@ -109,16 +117,20 @@ public class ValidationService implements Runnable {
 
     @SuppressWarnings("WeakerAccess")
     @UnitOfWork
-    public ValidationJob currentJob() {
+    public ValidationJob retrieveCurrentJob() {
+		logger.debug("Getting current job");
         return validationJobDAO.current();
     }
 
     @SuppressWarnings("WeakerAccess")
     @UnitOfWork
-    public void saveResult(ValidationJob job, VeraPDFValidationResult result) {
-        try {
-            if (!isAborted) {
-                DomainDocument document = job.getDocument();
+    public void saveResult(VeraPDFValidationResult result) {
+		boolean shouldCleanDB = false;
+    	try {
+            if (!currentJob.getStatus().equals(ValidationJob.Status.ABORTED)) {
+            	shouldCleanDB = true;
+				logger.debug("Saving validation job results");
+                DomainDocument document = currentJob.getDocument();
                 document.setBaseTestResult(result.getTestResult());
 
                 // Save errors where needed
@@ -128,28 +140,66 @@ public class ValidationService implements Runnable {
                 }
                 document.setValidationErrors(validationErrors);
 
-                // Link properties
-                document.setProperties(result.getProperties());
+                // Link properties and modification date
+				Map<String, String> properties = result.getProperties();
+				String modDateXMP = null;
+				String modDateInfoDict = null;
+				if (properties.containsKey(PROPERTY_NAME_MOD_DATE_XMP)) {
+					modDateXMP = properties.get(PROPERTY_NAME_MOD_DATE_XMP);
+					properties.remove(PROPERTY_NAME_MOD_DATE_XMP);
+				}
+				if (properties.containsKey(PROPERTY_NAME_MOD_DATE_INFO_DICT)) {
+					modDateInfoDict = properties.get(PROPERTY_NAME_MOD_DATE_INFO_DICT);
+					properties.remove(PROPERTY_NAME_MOD_DATE_INFO_DICT);
+				}
+				Date modDate = getModDate(modDateXMP, modDateInfoDict);
+				if (modDate != null) {
+					document.setLastModified(modDate);
+				}
+				document.setProperties(properties);
 
                 // And update document (note that document was detached from hibernate context, thus we need to save explicitly)
                 documentDAO.save(document);
-            }
+            } else {
+            	logger.debug("Validation job was aborted, don't save any results");
+			}
         } finally {
-            cleanJob(job);
+            cleanJob(currentJob, shouldCleanDB);
         }
     }
 
-    @SuppressWarnings("WeakerAccess")
-    @UnitOfWork
-    public void cleanJob(ValidationJob job) {
-        if (job == null) {
-            return;
-        }
-        if (job.getFilePath() != null) {
+    private static Date getModDate(String fromXMP, String fromInfoDict) {
+		if (fromXMP != null) {
+			try {
+				XMPDateTime fromISO8601 = XMPDateTimeFactory.createFromISO8601(fromXMP);
+				return fromISO8601.getCalendar().getTime();
+			} catch (XMPException e) {
+				return null;
+			}
+		} else if (fromInfoDict != null) {
+			try {
+				XMLGregorianCalendar xmlGregorianCalendar = DatatypeFactory.newInstance().newXMLGregorianCalendar(fromInfoDict);
+				return xmlGregorianCalendar.toGregorianCalendar().getTime();
+			} catch (DatatypeConfigurationException e) {
+				return null;
+			}
+		} else {
+			return null;
+		}
+	}
+
+    private void cleanJob(ValidationJob job, boolean shouldCleanDB) {
+    	logger.debug("Cleanup validation job");
+		if (job == null) {
+			return;
+		}
+    	if (job.getFilePath() != null) {
             if (!new File(job.getFilePath()).delete()) {
                 logger.warn("Failed to clean validation job file " + job.getFilePath());
             }
         }
-        validationJobDAO.remove(job);
+        if (shouldCleanDB) {
+			validationJobDAO.remove(job);
+		}
     }
 }
